breakpoints api

+++ src/include/registers.h
@@ -3,8 +3,8 @@
 
 #include <sys/types.h>
 
-long getpc(pid_t pid);
-void set_ip (pid_t pid, long ip);
-void dump_registers(pid_t pid);
+long long 	getpc(pid_t pid);
+void 		set_ip (pid_t pid, long long ip);
+void 		dump_registers(pid_t pid);
 
 #endif
